changed spacing in lists, fonts for titles etc
The titles might have to be changed back, but I'm adapting some
beautiful stuff from Luis that was originally requested by Stan.  Right
now we have sf fonts in most places, but not in table of contents.  And
the chapter numbering shows up even in chapter* places -- the part
summaries, for example.