More createspace fixes.
Changed fonts in figures to Times, so that they could be embedded.
Converted the Stern logo PNG>EPS>PDF.  This fixed the layers issue. I
believe the Stern logo will still be problematic because it  is not of
high enough DPI.  This only generates a warning at createspace that the
image may be blurry.  I'm going to give it a try anyway. God have mercy
on my soul.